CLDSRV-805: Regroup object related tests
Regroup copy, delete, get, head, put into an object.js file

Using 1 bucket on test suite to reduce gcp quota pressure
