--- tests/functional/raw-node/test/GCP/object/copy.js
@@ -1,157 +1,0 @@
-const assert = require('assert');
-const async = require('async');
-const arsenal = require('arsenal');
-const { GCP } = arsenal.storage.data.external.GCP;
-const { genUniqID, genBucketName, gcpRetry } = require('../../../utils/gcpUtils');
-const { getRealAwsConfig } =
-    require('../../../../aws-node-sdk/test/support/awsConfig');
-const {
-    CreateBucketCommand,
-    DeleteBucketCommand,
-    PutObjectCommand,
-} = require('@aws-sdk/client-s3');
-
-const credentialOne = 'gcpbackend';
-const bucketName = genBucketName('copy');
-
-describe('GCP: COPY Object', function testSuite() {
-    this.timeout(180000);
-    const config = getRealAwsConfig(credentialOne);
-    const gcpClient = new GCP(config);
-
-    before(async () => {
-        await gcpRetry(
-            gcpClient,
-            new CreateBucketCommand({ Bucket: bucketName }),
-        );
-    });
-
-    after(async () => {
-        await gcpRetry(
-            gcpClient,
-            new DeleteBucketCommand({ Bucket: bucketName }),
-        );
-    });
-
-    describe('without existing object in bucket', () => {
-        it('should return 404 and \'NoSuchKey\'', done => {
-            const missingObject = `nonexistingkey-${genUniqID()}`;
-            const someKey = `somekey-${genUniqID()}`;
-            gcpClient.copyObject({
-                Bucket: bucketName,
-                Key: someKey,
-                CopySource: `/${bucketName}/${missingObject}`,
-            }, err => {
-                assert(err);
-                assert.strictEqual(err.$metadata.httpStatusCode, 404);
-                assert.strictEqual(err.name, 'NoSuchKey');
-                return done();
-            });
-        });
-    });
-
-    describe('with existing object in bucket', () => {
-        beforeEach(async function beforeFn() {
-            this.currentTest.key = `somekey-${genUniqID()}`;
-            this.currentTest.copyKey = `copykey-${genUniqID()}`;
-            this.currentTest.initValue = `${genUniqID()}`;
-            const cmd = new PutObjectCommand({
-                Bucket: bucketName,
-                Key: this.currentTest.copyKey,
-                Metadata: {
-                    value: this.currentTest.initValue,
-                },
-            });
-            const res = await gcpClient.send(cmd);
-            this.currentTest.ETag = res.ETag;
-        });
-
-        afterEach(function afterFn(done) {
-            async.parallel([
-                next => gcpClient.deleteObject({
-                    Bucket: bucketName,
-                    Key: this.currentTest.key,
-                }, err => {
-                    if (err) {
-                        process.stdout.write(`err in deleting object ${err}\n`);
-                    }
-                    return next(err);
-                }),
-                next => gcpClient.deleteObject({
-                    Bucket: bucketName,
-                    Key: this.currentTest.copyKey,
-                }, err => {
-                    if (err) {
-                        process.stdout
-                            .write(`err in deleting copy object ${err}\n`);
-                    }
-                    return next(err);
-                }),
-            ], done);
-        });
-
-        it('should successfully copy with REPLACE directive',
-        function testFn(done) {
-            const newValue = `${genUniqID()}`;
-            async.waterfall([
-                next => gcpClient.copyObject({
-                    Bucket: bucketName,
-                    Key: this.test.key,
-                    CopySource: `/${bucketName}/${this.test.copyKey}`,
-                    MetadataDirective: 'REPLACE',
-                    Metadata: {
-                        value: newValue,
-                    },
-                }, err => {
-                    assert.equal(err, null,
-                        `Expected success, but got error ${err}`);
-                    return next();
-                }),
-                next => gcpClient.headObject({
-                    Bucket: bucketName,
-                    Key: this.test.key,
-                }, (err, res) => {
-                    if (err) {
-                        process.stdout
-                            .write(`err in retrieving object ${err}\n`);
-                        return next(err);
-                    }
-                    assert.strictEqual(res.ETag, this.test.ETag);
-                    assert.notStrictEqual(res.Metadata.value,
-                        this.test.initValue);
-                    return next();
-                }),
-            ], done);
-        });
-
-        it('should successfully copy with COPY directive',
-        function testFn(done) {
-            async.waterfall([
-                next => gcpClient.copyObject({
-                    Bucket: bucketName,
-                    Key: this.test.key,
-                    CopySource: `/${bucketName}/${this.test.copyKey}`,
-                    MetadataDirective: 'COPY',
-                }, err => {
-                    assert.equal(err, null,
-                        `Expected success, but got error ${err}`);
-                    return next();
-                }),
-                next => gcpClient.headObject({
-                    Bucket: bucketName,
-                    Key: this.test.key,
-                }, (err, res) => {
-                    if (err) {
-                        process.stdout
-                            .write(`err in retrieving object ${err}\n`);
-                        return next(err);
-                    }
-                    assert.strictEqual(res.ETag, this.test.ETag);
-                    assert.strictEqual(res.Metadata.value,
-                        this.test.initValue);
-                    return next();
-                }),
-            ], done);
-        });
-    });
-});
